# backend/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for external devices to access the internal Smartify system",
        "title": "Smartify Backend API",
        "contact": {
            "name": "Smartify Working Mail",
            "email": "smartifyprojectapp@gmail.com"
        },
        "version": "1.1"
    },
    "host": "213.226.112.206:22025",
    "basePath": "/api",
    "paths": {
        "/add_tutor": {
            "post": {
                "description": "Available only to authenticated users with the tutor role. Updates or creates a tutor record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tutor"
                ],
                "summary": "Adding/updating tutor information",
                "parameters": [
                    {
                        "description": "Tutor data to be updated",
                        "name": "tutor_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Tutor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful data update",
                        "schema": {
                            "$ref": "#/definitions/api.Tutor_succes"
                        }
                    },
                    "400": {
                        "description": "Invalid data or JSON",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "403": {
                        "description": "The user is not a tutor",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "500": {
                        "description": "Server error (database, etc.)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
        "/get_tutor": {
            "get": {
                "description": "Returns complete information about the current authenticated tutor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tutor"
                ],
                "summary": "Getting information about the tutor",
                "responses": {
                    "200": {
                        "description": "Tutor data",
                        "schema": {
                            "$ref": "#/definitions/database.Tutor"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "403": {
                        "description": "The user is not a tutor",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "500": {
                        "description": "Server error (database, etc.)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
        "/gettrackers": {
            "post": {
                "description": "Returns a list of trackers for an authenticated user. Valid access token is required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trackers"
                ],
                "summary": "Retrieving user trackers",
                "parameters": [
                    {
                        "description": "Request with access token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Get_trackers_request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with trackers",
                        "schema": {
                            "$ref": "#/definitions/api.Trackers"
                        }
                    },
                    "304": {
                        "description": "The data has not been modified (Not Modified)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized access (invalid token)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
        "/savetrackers": {
            "post": {
                "description": "Saves the user's trackers on the server for synchronization between devices. Requires valid access token and correct timestamp.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trackers"
                ],
                "summary": "Saving user trackers",
                "parameters": [
                    {
                        "description": "Data to be saved (token, trackers and timestamp)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Tracker_save"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trackers successfully saved",
                        "schema": {
                            "$ref": "#/definitions/api.Success_answer"
                        }
                    },
                    "304": {
                        "description": "The data has not been modified (Not Modified)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "400": {
                        "description": "Invalid request (incorrect data or time format)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized access (invalid token)",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/api.Error_answer"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Error_answer": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "error": {
                    "description": "Error message",
                    "type": "string"
                }
            }
        },
        "api.Get_trackers_request": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Authentication token (Access Token)",
                    "type": "string"
                }
            }
        },
...
        "api.Success_answer": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code",
                    "type": "integer"
                },
                "status": {
                    "description": "Status message",
                    "type": "string"
                }
            }
        },
...
        "api.Tracker_save": {
            "type": "object",
            "properties": {
                "timestamp": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "trackers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[tracker1",
                        " tracker2]"
                    ]
                }
            }
        },
        "api.Trackers": {
            "type": "object",
            "properties": {
                "trackers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[tracker1",
                        " tracker2]"
                    ]
                }
            }
        },
        "api.Tutor_succes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "database.Tutor": {
            "type": "object",
            "properties": {
                "cource": {
                    "type": "integer"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timestamp": {
                    "type": "string"
                },
                "university": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}